+++ src/main/java/net/jasper/mod/util/ClientHelpers.java
@@ -4,6 +4,7 @@ import net.jasper.mod.automation.PlayerRecorder;
 import net.jasper.mod.gui.option.PlayerAutomaOptionsScreen;
 import net.jasper.mod.util.data.LookingDirection;
 import net.jasper.mod.util.data.SlotClick;
+import net.jasper.mod.util.data.StartingPositionOffset;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.text.Text;
@@ -24,7 +25,7 @@ public class ClientHelpers {
         return scale;
     }
 
-    public static void centerPlayer() {
+    public static void positionPlayer() {
         // Center Camera
         PlayerEntity player= MinecraftClient.getInstance().player;
         assert player != null;
@@ -37,10 +38,14 @@ public class ClientHelpers {
             player.setYaw(dir.yaw());
         }
 
-        // Center player on current block
-        Vec3d playerPos = player.getPos();
-        BlockPos blockPos = new BlockPos((int) Math.floor(playerPos.x), (int) Math.floor(playerPos.y) - 1, (int) Math.floor(playerPos.z));
-        player.setPosition(blockPos.getX() + 0.5, blockPos.getY() + 1.0, blockPos.getZ() + 0.5);
+        // Center player on current block only if enabled in options
+        if (PlayerAutomaOptionsScreen.useDefaultStartingPositionOption.getValue()) {
+            StartingPositionOffset offset = PlayerAutomaOptionsScreen.setDefaultStartingPositionOption.getValue().getOffset();
+
+            Vec3d playerPos = player.getPos();
+            BlockPos blockPos = new BlockPos((int) Math.floor(playerPos.x), (int) Math.floor(playerPos.y), (int) Math.floor(playerPos.z));
+            player.setPosition(blockPos.getX() + offset.x(), blockPos.getY(), blockPos.getZ() + offset.z());
+        }
     }